feat(main): 更改类型为s32并添加新头文件 feat(util): 更新is_string和PrintOne结构以支持可变参数 feat(deque): 更改deque相关函数参数类型为s32 feat(functional): 更新hash函数支持s32类型 feat(hashtable): 更改hashtable相关结构参数类型为s32 feat(iterator): 更新iterator相关函数支持s32类型 feat(list): 更改list相关函数神元类型为s32 feat(map): 更新map类型为s32并添加测试支持 feat(multiset): 更新multiset类型为s32 feat(priority_queue): 更新priority_queue类型为s32 feat(set): 更新set类型为s32 feat(stack): 更新stack类型为s32 feat(unordered_map): 更新unordered_map类型为s32 feat(unordered_multiset): 更新unordered_multiset类型为s32 feat(test): 更改测试中涉及的所有int为s32并修复相关逻辑 feat(vector): 更改vector相关函数类型为s32并修复测试代码

+++ DwtSTL/functional.h
@@ -216,21 +216,21 @@ dwt_stl_TRIVIAL_HASH_FCN(char16_t)
 
 dwt_stl_TRIVIAL_HASH_FCN(char32_t)
 
-dwt_stl_TRIVIAL_HASH_FCN(short)
+dwt_stl_TRIVIAL_HASH_FCN(s16)
 
-dwt_stl_TRIVIAL_HASH_FCN(unsigned short)
+dwt_stl_TRIVIAL_HASH_FCN(u16)
 
-dwt_stl_TRIVIAL_HASH_FCN(int)
+dwt_stl_TRIVIAL_HASH_FCN(s32)
 
-dwt_stl_TRIVIAL_HASH_FCN(unsigned int)
+dwt_stl_TRIVIAL_HASH_FCN(u32)
 
 dwt_stl_TRIVIAL_HASH_FCN(long)
 
 dwt_stl_TRIVIAL_HASH_FCN(unsigned long)
 
-dwt_stl_TRIVIAL_HASH_FCN(long long)
+dwt_stl_TRIVIAL_HASH_FCN(s64)
 
-dwt_stl_TRIVIAL_HASH_FCN(unsigned long long)
+dwt_stl_TRIVIAL_HASH_FCN(u64)
 
 #undef dwt_stl_TRIVIAL_HASH_FCN
